refactor 파일 이동 및 파일 이름 수정
