+++ src/main/java/cart/config/Config.java
@@ -1,6 +1,6 @@
 package cart.config;
 
-import cart.domain.point.PointPolicy;
+import cart.domain.point.PointPolicy10Percent;
 import cart.domain.point.PointPolicyStrategy;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,6 +10,6 @@ public class Config {
 
     @Bean
     public PointPolicyStrategy pointPolicyStrategy() {
-        return new PointPolicy();
+        return new PointPolicy10Percent();
     }
 }
