+++ src/main/java/cart/controller/CartItemApiController.java
@@ -1,10 +1,10 @@
-package cart.ui;
+package cart.controller;
 
-import cart.application.CartItemService;
+import cart.service.CartItemService;
 import cart.domain.Member;
-import cart.dto.CartItemQuantityUpdateRequest;
-import cart.dto.CartItemRequest;
-import cart.dto.CartItemResponse;
+import cart.dto.cart.CartItemQuantityUpdateRequest;
+import cart.dto.cart.CartItemRequest;
+import cart.dto.cart.CartItemResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
