+++ src/main/java/cart/controller/ProductApiController.java
@@ -1,8 +1,8 @@
-package cart.ui;
+package cart.controller;
 
-import cart.application.ProductService;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.service.ProductService;
+import cart.dto.product.ProductRequest;
+import cart.dto.product.ProductResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
