+++ src/main/java/cart/dto/cart/CartItemQuantityUpdateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.cart;
 
 public class CartItemQuantityUpdateRequest {
     private Long quantity;
