+++ src/main/java/cart/dto/order/OrderResponse.java
@@ -1,7 +1,7 @@
 package cart.dto.order;
 
 import cart.domain.order.Order;
-import cart.dto.ProductResponse;
+import cart.dto.product.ProductResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
