+++ src/main/java/cart/repository/OrderRepository.java
@@ -10,8 +10,8 @@ import cart.dao.entity.ShippingDiscountPolicyEntity;
 import cart.dao.entity.ShippingFeeEntity;
 import cart.domain.Member;
 import cart.domain.point.Point;
-import cart.domain.ShippingDiscountPolicy;
-import cart.domain.ShippingFee;
+import cart.domain.shipping.ShippingDiscountPolicy;
+import cart.domain.shipping.ShippingFee;
 import cart.domain.order.Order;
 import cart.domain.order.OrderItem;
 import org.springframework.stereotype.Repository;
