+++ src/main/java/cart/service/OrderPolicyService.java
@@ -1,11 +1,13 @@
-package cart.application;
+package cart.service;
 
 import cart.domain.point.PointPolicyStrategy;
 import cart.dto.orderpolicy.OrderPolicyResponse;
 import cart.repository.OrderPolicyRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Transactional(readOnly = true)
 public class OrderPolicyService {
 
     private final OrderPolicyRepository orderPolicyRepository;
