+++ src/main/java/cart/service/ProductService.java
@@ -1,9 +1,9 @@
-package cart.application;
+package cart.service;
 
 import cart.domain.Product;
 import cart.dao.ProductDao;
-import cart.dto.ProductRequest;
-import cart.dto.ProductResponse;
+import cart.dto.product.ProductRequest;
+import cart.dto.product.ProductResponse;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
